토큰으로 유저 식별 및 API 요청마다 토큰 추가

+++ src/main/java/visang/showcase/aibackend/dto/request/triton/KnowledgeReqObject.java
@@ -9,5 +9,5 @@ public class KnowledgeReqObject {
     private String name;
     private String datatype;
     private List<Integer> shape;
-    private List<List<Integer>> data;
+    private List<List<Long>> data;
 }
